+++ packages/dartvel_cli/lib/dartvel_impl.dart
@@ -13,6 +13,7 @@ import 'src/commands/artifact_command.dart';
 import 'src/commands/capture_command.dart';
 import 'src/commands/webos_command.dart';
 import 'src/commands/explain_command.dart';
+import 'src/commands/flags_command.dart';
 import 'src/commands/spec_command.dart';
 import 'src/commands/engine_command.dart';
 import 'src/commands/ensure_path_command.dart';
@@ -60,6 +61,7 @@ Future<void> main(List<String> args) async {
     ..addCommand(CaptureCommand())
     ..addCommand(WebosCommand())
     ..addCommand(SpecCommand())
+    ..addCommand(FlagsCommand())
     ..addCommand(ExplainCommand())
     ..addCommand(EngineCommand())
     ..addCommand(EnsurePathCommand())